Implement S3 streaming proxy endpoint
The Bucket class supports sending a GET request directory to get a
response, or streaming the response. The request is authenticated using
the AWS Signing Signature v4 spec. Then the root external endpoint uses
Bucket to stream objects in its response, thereby creating a proxy to a
secure S3 bucket.