Demonstrate GitHub OAuth
We're using authlib via their starlette integration to implement a
GitHub OAuth client.

- GitHubOAuth configures/registers the client, and
github_oauth_dependency provides an instance of that GitHub client as a
path operation depenency.

- The GitHub access token is stored in the Cookie session.

Note that the endpoints for logging in and logging out are just stubs;
we'll develop those further so they're configured more naturally and
redirect to the right place.